requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.5.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.0
httpx==0.27.0
google-generativeai==0.3.1
pdfplumber==0.10.3
langchain==0.1.1
numpy==1.26.0
scipy==1.12.0
Jinja2==3.1.3
starlette==0.36.1
bcrypt==4.1.2
passlib==1.7.4
PyJWT==2.8.0
python-jose==3.3.0
